Add interactive functionalities
