+++ src/index.html
@@ -20,7 +20,7 @@
     </div>
     <hr>
     <div id="todo-list"></div>
-    <p class="clear">Clear all completed</p>
+    <button class="clear">Clear all completed</button>
   </div>
   <script src="https://kit.fontawesome.com/89df72a5a7.js" crossorigin="anonymous"></script>
 </body>
